Added Images to landing and navbar
Added stock images to the landing page and a section explaining the base cost of our services might want to change it to a destination fee and a flat rate for each day + equipment. Who knows. I also added navbar image matching the also added favicon. I'm enjoying making this.